python-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
rembg==2.0.56
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
